src/cloud_functions/bredg-web-reader-fn/requirements.txt
functions-framework==3.*
google-cloud-storage
aiohttp==3.9.*